samplesite/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2.7
django-bootstrap4==22.2
django-debug-toolbar==3.2.2
django-environ==0.9.0
django-filter==22.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
idna==3.3
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
